Modified REST API to not work from config.json and fixed errors in upload and delete for string parsing and empty file checking

+++ backend/uploader/uploadAvatar.go
@@ -2,17 +2,20 @@ package uploader
 
 import (
 	"context"
+	"log"
 
 	"github.com/sea350/ustart_micro/backend/uploader/uploaderpb"
 )
 
 //UploadAvatar uploads an avatar picture while returning the image link and error by first replacing the old image and then uploading a new image
 func (uploader *Uploader) UploadAvatar(ctx context.Context, req *uploaderpb.UploadAvatarRequest) (*uploaderpb.UploadAvatarResponse, error) {
-	err := uploader.Delete(ctx, req.OldImageLink)
-
-	if err != nil {
-		return nil, err
+	if req.OldImageLink != "" {
+		err := uploader.Delete(ctx, req.OldImageLink)
+		if err != nil {
+			return nil, err
+		}
 	}
+	log.Println("Hi")
 
 	newLink, err := uploader.Upload(ctx, req.Base64, req.UUID)
